refactor: merge file panel to attachment button
