+++ src/renderer/src/pages/home/Inputbar/AttachmentButton.tsx
@@ -1,12 +1,16 @@
-import { FileType } from '@renderer/types'
-import { filterSupportedFiles } from '@renderer/utils/file'
+import { QuickPanelReservedSymbol, useQuickPanel } from '@renderer/components/QuickPanel'
+import { useKnowledgeBases } from '@renderer/hooks/useKnowledge'
+import { FileType, KnowledgeBase, KnowledgeItem } from '@renderer/types'
+import { filterSupportedFiles, formatFileSize } from '@renderer/utils/file'
 import { Tooltip } from 'antd'
-import { Paperclip } from 'lucide-react'
-import { FC, useCallback, useImperativeHandle, useState } from 'react'
+import dayjs from 'dayjs'
+import { FileSearch, FileText, Paperclip, Upload } from 'lucide-react'
+import { Dispatch, FC, SetStateAction, useCallback, useImperativeHandle, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
 export interface AttachmentButtonRef {
   openQuickPanel: () => void
+  openFileSelectDialog: () => void
 }
 
 interface Props {
@@ -14,7 +18,7 @@ interface Props {
   couldAddImageFile: boolean
   extensions: string[]
   files: FileType[]
-  setFiles: (files: FileType[]) => void
+  setFiles: Dispatch<SetStateAction<FileType[]>>
   ToolbarButton: any
   disabled?: boolean
 }
@@ -29,9 +33,11 @@ const AttachmentButton: FC<Props> = ({
   disabled
 }) => {
   const { t } = useTranslation()
+  const quickPanel = useQuickPanel()
+  const { bases: knowledgeBases } = useKnowledgeBases()
   const [selecting, setSelecting] = useState<boolean>(false)
 
-  const onSelectFile = useCallback(async () => {
+  const openFileSelectDialog = useCallback(async () => {
     if (selecting) {
       return
     }
@@ -70,21 +76,86 @@ const AttachmentButton: FC<Props> = ({
     }
   }, [extensions, files, selecting, setFiles, t])
 
+  const openKnowledgeFileList = useCallback(
+    (base: KnowledgeBase) => {
+      quickPanel.open({
+        title: base.name,
+        list: base.items
+          .filter((file): file is KnowledgeItem => ['file'].includes(file.type))
+          .map((file) => {
+            const fileContent = file.content as FileType
+            return {
+              label: fileContent.origin_name || fileContent.name,
+              description:
+                formatFileSize(fileContent.size) + ' · ' + dayjs(fileContent.created_at).format('YYYY-MM-DD HH:mm'),
+              icon: <FileText />,
+              isSelected: files.some((f) => f.path === fileContent.path),
+              action: async ({ item }) => {
+                item.isSelected = !item.isSelected
+                if (fileContent.path) {
+                  setFiles((prevFiles) => {
+                    const fileExists = prevFiles.some((f) => f.path === fileContent.path)
+                    if (fileExists) {
+                      return prevFiles.filter((f) => f.path !== fileContent.path)
+                    } else {
+                      return fileContent ? [...prevFiles, fileContent] : prevFiles
+                    }
+                  })
+                }
+              }
+            }
+          }),
+        symbol: QuickPanelReservedSymbol.File,
+        multiple: true
+      })
+    },
+    [files, quickPanel, setFiles]
+  )
+
+  const items = useMemo(() => {
+    return [
+      {
+        label: t('chat.input.upload.upload_from_local'),
+        description: '',
+        icon: <Upload />,
+        action: () => openFileSelectDialog()
+      },
+      ...knowledgeBases.map((base) => {
+        const length = base.items?.filter(
+          (item): item is KnowledgeItem => ['file', 'note'].includes(item.type) && typeof item.content !== 'string'
+        ).length
+        return {
+          label: base.name,
+          description: `${length} ${t('files.count')}`,
+          icon: <FileSearch />,
+          disabled: length === 0,
+          isMenu: true,
+          action: () => openKnowledgeFileList(base)
+        }
+      })
+    ]
+  }, [knowledgeBases, openFileSelectDialog, openKnowledgeFileList, t])
+
   const openQuickPanel = useCallback(() => {
-    onSelectFile()
-  }, [onSelectFile])
+    quickPanel.open({
+      title: t('chat.input.upload.attachment'),
+      list: items,
+      symbol: QuickPanelReservedSymbol.File
+    })
+  }, [items, quickPanel, t])
 
   useImperativeHandle(ref, () => ({
-    openQuickPanel
+    openQuickPanel,
+    openFileSelectDialog
   }))
 
   return (
     <Tooltip
       placement="top"
-      title={couldAddImageFile ? t('chat.input.upload.label') : t('chat.input.upload.document')}
+      title={couldAddImageFile ? t('chat.input.upload.image_or_document') : t('chat.input.upload.document')}
       mouseLeaveDelay={0}
       arrow>
-      <ToolbarButton type="text" onClick={onSelectFile} disabled={disabled}>
+      <ToolbarButton type="text" onClick={openFileSelectDialog} disabled={disabled}>
         <Paperclip size={18} style={{ color: files.length ? 'var(--color-primary)' : 'var(--color-icon)' }} />
       </ToolbarButton>
     </Tooltip>
